% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLinearSquare_Functions.R
\name{Find_Wmatrix_MeanL2Error}
\alias{Find_Wmatrix_MeanL2Error}
\title{Find_Wmatrix_MeanL2Error}
\usage{
Find_Wmatrix_MeanL2Error(TestingData, TestingLables, W.Matrix,
  BinaryClassification)
}
\arguments{
\item{W.Matrix}{Weight values representing the Linear funciton [p+1,Iterations],}

\item{BinaryClassification}{binary or integer value that determines how the algorithm classifies the retulsts
If (BinaryClassification):
 Then Classification is either 0 or 1.}

\item{TrainingData}{numeric imput feature matrix [n x p]}

\item{TrainingLabels}{numberic input label vector [n]
either all 0/1 for binary classification or other real numbers for regression}

\item{StepSize.Scalar}{scalar integer, determines the size of each step.}
}
\value{
L2Error.Matrix
}
\description{
Makes iterative steps using gradient decent to find a solution to the Linear Models problem
}
\examples{

}
