% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLogistic_Functions.R
\name{LMLogisticLossEarlyStoppingCV}
\alias{LMLogisticLossEarlyStoppingCV}
\title{Logistic Models LogisticLossIterations}
\usage{
LMLogisticLossEarlyStoppingCV(TrainingData, TrainingLabels, fold.vec,
  folds.n = 4, max.iterations = 30)
}
\arguments{
\item{TrainingData}{numeric imput feature matrix [n x p]}

\item{TrainingLabels}{numberic input label vector [n],
either all 0/1 for binary classification or other real numbers for regression}

\item{max.iterations}{integer that determines the number of steps taken to find the optimal}
}
\value{
List of :

Returned Matrix is[n_features+1 x max.iterations], where the first element of the weight vector is be the intercept term(AKA Bias).
}
\description{
Makes iterative steps using gradient decent to find a solution to the Logistic Models problem
}
\examples{
## Example: 1 With The Spam DataSet:##

## Example: 2  With The ZipTrain Dataset:##

}
