% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLogistic_Functions.R
\name{LMLogisticLossL2}
\alias{LMLogisticLossL2}
\title{LMLogisticLossL2}
\usage{
LMLogisticLossL2(Normalized_TrainingData, TrainingLabels, penalty,
  opt.thresh, initial.weight.vec)
}
\arguments{
\item{Normalized_TrainingData}{numeric imput feature matrix [n x p]}

\item{TrainingLabels}{numberic input label vector [n],
either all 0/1 for binary classification or other real numbers for regression}

\item{initial.weight.vec}{}
}
\value{
numeric matrix of the weight matrix at each step, from 1 to Iterations.
Returned Matrix is[n_features+1 x max.iterations], where the first element of the weight vector is be the intercept term(AKA Bias).
}
\description{
Makes iterative steps using gradient decent to find a solution to the Logistic Models problem
}
\examples{
## Example: 1 With The Spam DataSet:##

## Example: 2  With The ZipTrain Dataset:##

}
