% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLinearSquare_Functions.R
\name{LMSquare_Gradient}
\alias{LMSquare_Gradient}
\title{LMSquare_Gradient}
\usage{
LMSquare_Gradient(TrainingData, TrainingLabels, W.Vector)
}
\arguments{
\item{TrainingData}{numeric imput feature matrix [n x p]}

\item{TrainingLabels}{numberic input label vector [n]}

\item{W.Vector}{Weight values representing the Linear funciton [p+1]}
}
\value{
returns the gradient of the W.Vector, in respect to the TrainingData, and TrainingLabels
Uses the following equation to find the gradients:
#Weight.Gradient = sum(W.Vector*TrainingData - TrainingLabels)*TrainingLabels
#BiasGradient = sum(Yb.hat - TrainingLabels)
}
\description{
LMSquare_Gradient
}
