% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLinearSquare_Functions.R
\name{LMSquareLossEarlyStoppingCV}
\alias{LMSquareLossEarlyStoppingCV}
\title{Linear Models SquareLossIterations}
\usage{
LMSquareLossEarlyStoppingCV(TrainingData, TrainingLabels, fold.vec,
  folds.n = 4, max.iterations = 30)
}
\arguments{
\item{TrainingData}{numeric imput feature matrix [n x p]}

\item{TrainingLabels}{numberic input label vector [n],
either all 0/1 for binary classification or other real numbers for regression}

\item{fold.vec}{vector indicating the fold each element is in [n]}

\item{max.iterations}{integer that determines the number of steps taken to find the optimal}

\item{fold.n}{the number of folds that are to be preformed}
}
\value{
List of :

Returned Matrix is[n_features+1 x max.iterations], where the first element of the weight vector is be the intercept term(AKA Bias).
}
\description{
Makes iterative steps using gradient decent to find a solution to the Linear Models problem
}
\examples{
## Example: 1 With The Spam DataSet:##
ES.List <-LMSquareLossEarlyStoppingCV(Spam$TrainingData, Spam$TrainingLabels,Spam$Folds.Vec,Spam$Folds.n,Spam$Iterations)
DeNormalizedWeights <- ES.List$w.mat

DeNorm.Error <-Find_Wmatrix_MeanL1Error(Spam$TrainingData, Spam$TrainingLabels,(DeNormalizedWeights),Spam$BinaryClassification)
barplot(DeNorm.Error,main = "LMSquareLossEarlyStoppingCV:Spam",xlab = "Iteration",ylab = "Error",beside = TRUE)

## Example: 2  With The ZipTrain Dataset:##
}
