% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{NormalizeMatrix_List}
\alias{NormalizeMatrix_List}
\title{NormalizeMatrix_List}
\usage{
NormalizeMatrix_List(Matrix)
}
\arguments{
\item{Matrix}{[X x Y]}
}
\value{
Same as NormalizeMatrix, but returns contents within a list, so the standard deviation can be used later

Uses the formula: ((Matrix - Matrix.mean)/Matrix.sd)
}
\description{
NormalizeMatrix_List
}
\examples{
Norm_List = NormalizeMatrix_List(as.matrix(c(1,2,3)))
print(Norm_List)
# Returns:
$NormalizedMatrix
[,1]
[1,] -1.224745
[2,]  0.000000
[3,]  1.224745

$sd
[1] 0.8164966
}
