% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{NormalizeVector}
\alias{NormalizeVector}
\title{NormalizeVector}
\usage{
NormalizeVector(Vector)
}
\arguments{
\item{Vector}{[X]}
}
\value{
takes the absolute sum making the total distance, and divides by the length

Uses the formula: ((Matrix - Matrix.mean)/Matrix.sd)
}
\description{
NormalizeVector
}
\examples{
Norm_List = NormalizeMatrix_List(as.matrix(c(1,2,3)))
print(Norm_List)
# Returns:
$NormalizedMatrix
[,1]
[1,] -1.224745
[2,]  0.000000
[3,]  1.224745

$sd
[1] 0.8164966
}
