% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Random_Folds}
\alias{Random_Folds}
\title{Random_Folds
  Simple Function that creates a list of integers that indicates the fold Number each element is apart of.
I included error checking to prevent any incorrect parameters.}
\usage{
Random_Folds(Size, Folds)
}
\arguments{
\item{Size}{is the length of the array to be produced []}

\item{Folds}{the number of numbers to be produced with the range(1,Folds)}
}
\value{
list of [Size] length with random integers with the range(1,Folds)
}
\description{
Random_Folds
  Simple Function that creates a list of integers that indicates the fold Number each element is apart of.
I included error checking to prevent any incorrect parameters.
}
\examples{
 ~~ Example: 1 ~~
Folds =  Random_Folds(10,2)
print(Folds)
Returns:
[1] 1 1 2 2 2 1 1 1 2 1

 ~~ Example: 2 ~~
 Folds =  Random_Folds(10,5)
 print(Folds)
 [1] 4 3 1 5 5 1 5 3 1 5

}
